// src/payload-types.ts
export interface Show {
  id: number;
  name: 'The Court' | 'Fit School' | 'Maskani' | 'The Gameweek';
  description?: string | null;
  cadence?: string | null;
  playlistId?: string | null;
  mode?: ('day' | 'night') | null;
  updatedAt: string;
  createdAt: string;
}